extract i18next backend initializer to separate file
